1.3.14 Release

See merge request darkanrs/world-server!558

+++ src/main/java/com/rs/game/content/pets/PetManager.java
@@ -16,9 +16,6 @@
 //
 package com.rs.game.content.pets;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import com.rs.cache.loaders.ItemDefinitions;
 import com.rs.cache.loaders.NPCDefinitions;
 import com.rs.game.content.ItemConstants;
@@ -27,6 +24,9 @@ import com.rs.game.model.entity.player.Player;
 import com.rs.lib.game.Animation;
 import com.rs.lib.game.Item;
 import com.rs.lib.game.Tile;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * The pet manager.
@@ -164,13 +164,18 @@ public final class PetManager {
 				player.sendMessage("Your troll baby won't eat this item.");
 				return;
 			}
-			if (trollBabyName == null) {
-				trollBabyName = ItemDefinitions.getDefs(foodId).getName();
-				npc.setName(trollBabyName);
-				npc.setNextForceTalk(new ForceTalk("YUM! Me likes " + trollBabyName + "!"));
-			}
-			player.getInventory().deleteItem(foodId, 1);
-			player.sendMessage("Your pet happily eats the " + ItemDefinitions.getDefs(foodId).getName() + ".");
+			player.sendOptionDialogue("Are you sure you want to feed your " + ItemDefinitions.getDefs(foodId).getName() + " to the troll?", ops -> {
+				ops.add("Yes, I am sure. I know I won't get it back.", () -> {
+					if (!player.getInventory().containsItem(foodId))
+						return;
+					trollBabyName = ItemDefinitions.getDefs(foodId).getName();
+					npc.setName(trollBabyName);
+					npc.setNextForceTalk(new ForceTalk("YUM! Me likes " + trollBabyName + "!"));
+					player.getInventory().deleteItem(foodId, 1);
+					player.sendMessage("Your pet happily eats the " + ItemDefinitions.getDefs(foodId).getName() + ".");
+				});
+				ops.add("Nevermind.");
+			});
 			return;
 		}
 		for (int food : pets.getFood())
